requirements.txt

pytesseract
Pillow